requirements.txt
FastAPI
pydantic
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
fasttext
